# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CourseProgressDTO:
    properties:
      completed_at:
        type: string
      course_id:
        type: integer
      id:
        type: integer
      score:
        type: integer
      started_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  dto.SectionProgressDTO:
    properties:
      completed_at:
        type: string
      course_id:
        type: integer
      id:
        type: integer
      progress:
        type: number
      section_id:
        type: integer
      started_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  dto.UnitProgressDTO:
    properties:
      completed_at:
        type: string
      id:
        type: integer
      last_word_id:
        type: integer
      progress:
        type: number
      section_id:
        type: integer
      started_at:
        type: string
      status:
        type: string
      unit_id:
        type: integer
      updated_at:
        type: string
    type: object
  dto.WordCreateDTO:
    properties:
      examples:
        example:
        - Hello
        - ' world!'
        items:
          type: string
        type: array
      phonetic:
        example: həˈləʊ
        type: string
      tags:
        example:
        - common
        - greeting
        items:
          type: string
        type: array
      text:
        example: hello
        type: string
      translation:
        example: 你好
        type: string
    required:
    - text
    - translation
    type: object
  dto.WordResponseDTO:
    properties:
      created_at:
        example: "2025-01-26 18:00:00"
        type: string
      examples:
        example:
        - Hello
        - ' world!'
        items:
          type: string
        type: array
      id:
        example: 1
        type: integer
      phonetic:
        example: həˈləʊ
        type: string
      tags:
        example:
        - common
        - greeting
        items:
          type: string
        type: array
      text:
        example: hello
        type: string
      translation:
        example: 你好
        type: string
      updated_at:
        example: "2025-01-26 18:00:00"
        type: string
    type: object
  handler.ListResponse:
    properties:
      items:
        description: 列表数据
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  handler.Response:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:9000
info:
  contact:
    email: lazyjean@example.com
    name: LazyJean
  description: 生词本服务 API 文档
  title: 生词本 API
  version: "1.0"
paths:
  /learning/courses/{courseId}/progress:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CourseProgressDTO'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: 获取课程学习进度列表
      tags:
      - learning
    post:
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: integer
      - description: 学习状态
        enum:
        - not_started
        - in_progress
        - completed
        in: query
        name: status
        required: true
        type: string
      - description: 得分
        in: query
        name: score
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CourseProgressDTO'
              type: object
      security:
      - Bearer: []
      summary: 保存课程学习进度
      tags:
      - learning
  /learning/courses/{courseId}/sections/progress:
    get:
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SectionProgressDTO'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: 获取课程的章节学习进度列表
      tags:
      - learning
  /learning/sections/{sectionId}/progress:
    get:
      parameters:
      - description: 章节ID
        in: path
        name: sectionId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SectionProgressDTO'
              type: object
      security:
      - Bearer: []
      summary: 获取章节学习进度
      tags:
      - learning
    post:
      parameters:
      - description: 章节ID
        in: path
        name: sectionId
        required: true
        type: integer
      - description: 课程ID
        in: query
        name: courseId
        required: true
        type: integer
      - description: 学习状态
        enum:
        - not_started
        - in_progress
        - completed
        in: query
        name: status
        required: true
        type: string
      - description: 进度
        in: query
        maximum: 100
        minimum: 0
        name: progress
        required: true
        type: number
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SectionProgressDTO'
              type: object
      security:
      - Bearer: []
      summary: 保存章节学习进度
      tags:
      - learning
  /learning/sections/{sectionId}/units/progress:
    get:
      parameters:
      - description: 章节ID
        in: path
        name: sectionId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UnitProgressDTO'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: 获取章节的单元学习进度列表
      tags:
      - learning
  /learning/units/{unitId}/progress:
    get:
      parameters:
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UnitProgressDTO'
              type: object
      security:
      - Bearer: []
      summary: 获取单元学习进度
      tags:
      - learning
    post:
      parameters:
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: integer
      - description: 章节ID
        in: query
        name: sectionId
        required: true
        type: integer
      - description: 学习状态
        enum:
        - not_started
        - in_progress
        - completed
        in: query
        name: status
        required: true
        type: string
      - description: 进度
        in: query
        maximum: 100
        minimum: 0
        name: progress
        required: true
        type: number
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UnitProgressDTO'
              type: object
      security:
      - Bearer: []
      summary: 保存单元学习进度
      tags:
      - learning
  /v1/login:
    post:
      responses: {}
  /v1/register:
    post:
      responses: {}
  /v1/words:
    get:
      consumes:
      - application/json
      description: 分页获取单词列表
      parameters:
      - description: Bearer 令牌
        in: header
        name: Authorization
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/handler.ListResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/dto.WordResponseDTO'
                        type: array
                    type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 获取单词列表
      tags:
      - words
    post:
      consumes:
      - application/json
      description: 创建一个新的单词记录
      parameters:
      - description: Bearer 令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 单词信息
        in: body
        name: word
        required: true
        schema:
          $ref: '#/definitions/dto.WordCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.WordResponseDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 创建新单词
      tags:
      - words
  /v1/words/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的单词
      parameters:
      - description: Bearer 令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 单词ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 单词不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 删除单词
      tags:
      - words
    get:
      consumes:
      - application/json
      description: 通过ID获取单词的详细信息
      parameters:
      - description: Bearer 令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 单词ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.WordResponseDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 单词不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 获取单词详情
      tags:
      - words
  /v1/words/search:
    get:
      consumes:
      - application/json
      description: 根据条件搜索单词
      parameters:
      - description: Bearer 令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 单词文本
        in: query
        name: text
        type: string
      - collectionFormat: csv
        description: 标签列表
        in: query
        items:
          type: string
        name: tags
        type: array
      - description: 最小难度
        in: query
        name: minDifficulty
        type: integer
      - description: 最大难度
        in: query
        name: maxDifficulty
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/handler.ListResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/dto.WordResponseDTO'
                        type: array
                    type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - Bearer: []
      summary: 搜索单词
      tags:
      - words
schemes:
- http
securityDefinitions:
  Bearer:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"